+++ src/main/java/xyz/nikitacartes/easywhitelist/commands/EasyDeOpCommand.java
@@ -1,0 +1,22 @@
+package xyz.nikitacartes.easywhitelist.commands;
+
+import com.mojang.brigadier.CommandDispatcher;
+import net.minecraft.server.command.ServerCommandSource;
+
+import static com.mojang.brigadier.arguments.StringArgumentType.getString;
+import static com.mojang.brigadier.arguments.StringArgumentType.word;
+import static net.minecraft.server.command.CommandManager.argument;
+import static net.minecraft.server.command.CommandManager.literal;
+import static net.minecraft.server.dedicated.command.DeOpCommand.deop;
+import static xyz.nikitacartes.easywhitelist.EasyWhitelist.getProfileFromNickname;
+
+public class EasyDeOpCommand {
+
+    public static void registerCommand(CommandDispatcher<ServerCommandSource> dispatcher) {
+        dispatcher.register(literal("easydeop")
+                .requires(source -> source.hasPermissionLevel(3))
+                .then(argument("targets", word())
+                        .executes(ctx ->
+                                deop(ctx.getSource(), getProfileFromNickname(getString(ctx, "targets"))))));
+    }
+}