+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.2.0 Minecraft 1.17.1
+
+Add offline version of some commands:
+
+* `easywhitelist remove <nickname>`
+* `easyban <nickname>`
+* `easypardon <nickname>`
+* `easyop <nickname>`
+* `easydeop <nickname>`
+
 ### 0.1.0 Minecraft 1.17.1
 
 Add command `easywhitelist add <nickname>` that add offline uuid and nickname in whitelist